Fix #14763 - Include extension specific properties only - Refactor: Move extensions scanning to MainThreadExtensionService
